requirements.txt
# Core dependencies
streamlit>=1.30.0
python-dotenv>=1.0.0

# Document processing
pypdf>=4.0.0

# Embeddings and Vector Store
faiss-cpu>=1.7.4
sentence-transformers>=2.3.0

# LLM Integration
openai>=1.12.0
groq>=0.4.0

# Utilities
tiktoken>=0.5.0
tenacity>=8.2.0
loguru>=0.7.0
